change User & logon form

+++ src/main/java/ua/bu/entity/Role.java
@@ -1,8 +1,6 @@
 package ua.bu.entity;
 
 import javax.persistence.*;
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Table(name = "roles")
@@ -10,23 +8,17 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
     private String loginName;
     private String role;
 
-//    @OneToMany(cascade = CascadeType.ALL, fetch = FetchType.EAGER, targetEntity = User.class)
-//    @JoinColumn(name = "role_id")
-//    private List<User> users = new ArrayList<>();
-
-
-    public int getId() {
+    public long getId() {
         return id;
     }
 
     public void setId(int id) {
         this.id = id;
     }
-
 
 
     public String getRole() {
@@ -36,14 +28,7 @@ public class Role {
     public void setRole(String role) {
         this.role = role;
     }
-//
-//    public List<User> getUsers() {
-//        return users;
-//    }
-//
-//    public void setUsers(List<User> users) {
-//        this.users = users;
-//    }
+
 
     public String getLoginName() {
         return loginName;
@@ -59,7 +44,6 @@ public class Role {
                 "id=" + id +
                 ", loginName='" + loginName + '\'' +
                 ", role='" + role + '\'' +
-//                ", users=" + users +
                 '}';
     }
 }
